+++ R/tm_g_gh_boxplot.R
@@ -328,6 +328,16 @@ srv_g_boxplot <- function(input,
       anl_chunks()$ANL,
       facet_var,
       sprintf("Variable %s is not available in data %s", facet_var, dataname))
+
+    validate(need(
+      !facet_var %in% c("ACTARM","ARM")[!c("ACTARM","ARM") %in% trt_group],
+      sprintf("You can not choose %s as facetting variable for treatment variable %s.", facet_var, trt_group)
+      ))
+    validate(need(
+      !xaxis %in% c("ACTARM","ARM")[!c("ACTARM","ARM") %in% trt_group],
+      sprintf("You can not choose %s as facetting variable for treatment variable %s.", xaxis, trt_group)
+    ))
+
     chunks_push(
       chunks = private_chunks,
       id = "boxplot",