+++ R/tm_g_gh_correlationplot.R
@@ -348,9 +348,9 @@ srv_g_correlationplot <- function(input,
 
     if (!is.null(input$facet_var)) {
       validate(need(
-        identical(input$trt_group, input$facet_var),
-        "Please choose the same facetting variable as the treatment variable"
-        ))
+        !input$facet_var %in% c("ACTARM","ARM")[!c("ACTARM","ARM") %in% input$trt_group],
+        sprintf("You can not choose %s as facetting variable for treatment variable %s.", input$facet_var, input$trt_group)
+      ))
       }
     # analysis
     private_chunks <- chunks$new()